AgriPulse_Backend/requirements.txt
Django
django-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
requests